Move IClientSettingsManager down, and re-combine with IClientSettingsReader
