+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/LanguageClient/WorkspaceConfigurationChangedListener.cs
@@ -5,6 +5,7 @@ using System;
 using System.ComponentModel.Composition;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Razor.Workspaces.Settings;
 using Microsoft.VisualStudio.LanguageServer.ContainedLanguage;
 using Microsoft.VisualStudio.Razor.Settings;
