+++ src/Razor/src/Microsoft.VisualStudio.LegacyEditor.Razor/Settings/WorkspaceEditorSettings.cs
@@ -4,7 +4,7 @@
 using System;
 using System.ComponentModel.Composition;
 using Microsoft.CodeAnalysis.Razor.Settings;
-using Microsoft.VisualStudio.Razor.Settings;
+using Microsoft.CodeAnalysis.Razor.Workspaces.Settings;
 
 namespace Microsoft.VisualStudio.LegacyEditor.Razor.Settings;
